Support an optional user destination.
Allow the user to specify the desired final stop along the way
instead of forcing the train's own final stop. The output will
show the ETA and distance relative to the stop where the user
needs to get off. Even DB realized that this interesting and offers this
feature in the web UI.